+++ src/spring-cloud/azext_spring_cloud/custom.py
@@ -306,19 +306,21 @@ def app_update(cmd, client, resource_group, service, name,
                main_entry=None,
                env=None,
                enable_persistent_storage=None,
-               https_only=None):
+               https_only=None,
+               enable_end_to_end_tls=None):
     _check_active_deployment_exist(client, resource_group, service, name)
     resource = client.services.get(resource_group, service)
     location = resource.location
 
-    properties = models.AppResourceProperties(public=assign_endpoint, https_only=https_only)
+    properties = models_20201101preview.AppResourceProperties(public=assign_endpoint, https_only=https_only,
+                                                              enable_end_to_end_tls=enable_end_to_end_tls)
     if enable_persistent_storage is True:
         properties.persistent_disk = models.PersistentDisk(
             size_in_gb=_get_persistent_disk_size(resource.sku.tier), mount_path="/persistent")
     if enable_persistent_storage is False:
         properties.persistent_disk = models.PersistentDisk(size_in_gb=0)
 
-    app_resource = models.AppResource()
+    app_resource = models_20201101preview.AppResource()
     app_resource.properties = properties
     app_resource.location = location
 
@@ -1031,16 +1033,16 @@ def config_repo_list(cmd, client, resource_group, name):
 
 def binding_list(cmd, client, resource_group, service, app):
     _check_active_deployment_exist(client, resource_group, service, app)
-    return client.list(resource_group, service, app)
+    return client.bindings.list(resource_group, service, app)
 
 
 def binding_get(cmd, client, resource_group, service, app, name):
     _check_active_deployment_exist(client, resource_group, service, app)
-    return client.get(resource_group, service, app, name)
+    return client.bindings.get(resource_group, service, app, name)
 
 
 def binding_remove(cmd, client, resource_group, service, app, name):
-    return client.delete(resource_group, service, app, name)
+    return client.bindings.delete(resource_group, service, app, name)
 
 
 def binding_cosmos_add(cmd, client, resource_group, service, app, name,
@@ -1063,7 +1065,7 @@ def binding_cosmos_add(cmd, client, resource_group, service, app, name,
         binding_parameters['collectionName'] = collection_name
 
     try:
-        primary_key = _get_cosmosdb_primary_key(client, resource_id)
+        primary_key = _get_cosmosdb_primary_key(client.bindings, resource_id)
     except:
         raise CLIError(
             "Couldn't get cosmosdb {}'s primary key".format(resource_name))
@@ -1075,7 +1077,7 @@ def binding_cosmos_add(cmd, client, resource_group, service, app, name,
         key=primary_key,
         binding_parameters=binding_parameters
     )
-    return client.create_or_update(resource_group, service, app, name, properties)
+    return client.bindings.create_or_update(resource_group, service, app, name, properties)
 
 
 def binding_cosmos_update(cmd, client, resource_group, service, app, name,
@@ -1083,7 +1085,7 @@ def binding_cosmos_update(cmd, client, resource_group, service, app, name,
                           key_space=None,
                           collection_name=None):
     _check_active_deployment_exist(client, resource_group, service, app)
-    binding = client.get(resource_group, service, app, name).properties
+    binding = client.bindings.get(resource_group, service, app, name).properties
     resource_id = binding.resource_id
     resource_name = binding.resource_name
     binding_parameters = {}
@@ -1092,7 +1094,7 @@ def binding_cosmos_update(cmd, client, resource_group, service, app, name,
     binding_parameters['collectionName'] = collection_name
 
     try:
-        primary_key = _get_cosmosdb_primary_key(client, resource_id)
+        primary_key = _get_cosmosdb_primary_key(client.bindings, resource_id)
     except:
         raise CLIError(
             "Couldn't get cosmosdb {}'s primary key".format(resource_name))
@@ -1101,7 +1103,7 @@ def binding_cosmos_update(cmd, client, resource_group, service, app, name,
         key=primary_key,
         binding_parameters=binding_parameters
     )
-    return client.update(resource_group, service, app, name, properties)
+    return client.bindings.update(resource_group, service, app, name, properties)
 
 
 def binding_mysql_add(cmd, client, resource_group, service, app, name,
@@ -1124,7 +1126,7 @@ def binding_mysql_add(cmd, client, resource_group, service, app, name,
         key=key,
         binding_parameters=binding_parameters
     )
-    return client.create_or_update(resource_group, service, app, name, properties)
+    return client.bindings.create_or_update(resource_group, service, app, name, properties)
 
 
 def binding_mysql_update(cmd, client, resource_group, service, app, name,
@@ -1140,7 +1142,7 @@ def binding_mysql_update(cmd, client, resource_group, service, app, name,
         key=key,
         binding_parameters=binding_parameters
     )
-    return client.update(resource_group, service, app, name, properties)
+    return client.bindings.update(resource_group, service, app, name, properties)
 
 
 def binding_redis_add(cmd, client, resource_group, service, app, name,
@@ -1155,7 +1157,7 @@ def binding_redis_add(cmd, client, resource_group, service, app, name,
     binding_parameters['useSsl'] = use_ssl
     primary_key = None
     try:
-        primary_key = _get_redis_primary_key(client, resource_id)
+        primary_key = _get_redis_primary_key(client.bindings, resource_id)
     except:
         raise CLIError(
             "Couldn't get redis {}'s primary key".format(resource_name))
@@ -1168,13 +1170,13 @@ def binding_redis_add(cmd, client, resource_group, service, app, name,
         binding_parameters=binding_parameters
     )
 
-    return client.create_or_update(resource_group, service, app, name, properties)
+    return client.bindings.create_or_update(resource_group, service, app, name, properties)
 
 
 def binding_redis_update(cmd, client, resource_group, service, app, name,
                          disable_ssl=None):
     _check_active_deployment_exist(client, resource_group, service, app)
-    binding = client.get(resource_group, service, app, name).properties
+    binding = client.bindings.get(resource_group, service, app, name).properties
     resource_id = binding.resource_id
     resource_name = binding.resource_name
     binding_parameters = {}
@@ -1183,7 +1185,7 @@ def binding_redis_update(cmd, client, resource_group, service, app, name,
 
     primary_key = None
     try:
-        primary_key = _get_redis_primary_key(client, resource_id)
+        primary_key = _get_redis_primary_key(client.bindings, resource_id)
     except:
         raise CLIError(
             "Couldn't get redis {}'s primary key".format(resource_name))
@@ -1192,7 +1194,7 @@ def binding_redis_update(cmd, client, resource_group, service, app, name,
         key=primary_key,
         binding_parameters=binding_parameters
     )
-    return client.update(resource_group, service, app, name, properties)
+    return client.bindings.update(resource_group, service, app, name, properties)
 
 
 def _get_cosmosdb_primary_key(client, resource_id):
@@ -1382,7 +1384,8 @@ def certificate_remove(cmd, client, resource_group, service, name):
 
 def domain_bind(cmd, client, resource_group, service, app,
                 domain_name,
-                certificate=None):
+                certificate=None,
+                enable_end_to_end_tls=None):
     _check_active_deployment_exist(client, resource_group, service, app)
     properties = models.CustomDomainProperties()
     if certificate is not None:
@@ -1391,7 +1394,26 @@ def domain_bind(cmd, client, resource_group, service, app,
             thumbprint=certificate_response.properties.thumbprint,
             cert_name=certificate
         )
+    if enable_end_to_end_tls is not None:
+        _update_app_e2e_tls(cmd, resource_group, service, app, enable_end_to_end_tls)
+
     return client.custom_domains.create_or_update(resource_group, service, app, domain_name, properties)
+
+
+def _update_app_e2e_tls(cmd, resource_group, service, app, enable_end_to_end_tls):
+    client = get_mgmt_service_client(cmd.cli_ctx, AppPlatformManagementClient_20201101preview)
+    resource = client.services.get(resource_group, service)
+    location = resource.location
+
+    properties = models_20201101preview.AppResourceProperties(enable_end_to_end_tls=enable_end_to_end_tls)
+    app_resource = models_20201101preview.AppResource()
+    app_resource.properties = properties
+    app_resource.location = location
+
+    logger.warning("Set end to end tls for app '{}'".format(app))
+    poller = client.apps.update(
+        resource_group, service, app, app_resource)
+    return poller.result()
 
 
 def domain_show(cmd, client, resource_group, service, app, domain_name):
@@ -1406,7 +1428,8 @@ def domain_list(cmd, client, resource_group, service, app):
 
 def domain_update(cmd, client, resource_group, service, app,
                   domain_name,
-                  certificate=None):
+                  certificate=None,
+                  enable_end_to_end_tls=None):
     _check_active_deployment_exist(client, resource_group, service, app)
     properties = models.CustomDomainProperties()
     if certificate is not None:
@@ -1415,6 +1438,9 @@ def domain_update(cmd, client, resource_group, service, app,
             thumbprint=certificate_response.properties.thumbprint,
             cert_name=certificate
         )
+    if enable_end_to_end_tls is not None:
+        _update_app_e2e_tls(cmd, resource_group, service, app, enable_end_to_end_tls)
+
     return client.custom_domains.create_or_update(resource_group, service, app, domain_name, properties)